refactor: consolidate extension logic into core class and update configuration handling
- Fixed issue with timeouts not being removed before creating new ones.
- Fixed issue with spawning a process to launch the extension preferences instead of using the built-in method provided by the Extension base class.
- Moved core functionality from the extension class to a dedicated VSCodeWorkspacesCore class for better organization and maintainability.
- Updated the extension to utilize the new core class, simplifying the enable and disable methods.
- Introduced a new configuration file format for prettier settings, replacing the previous .prettierrc.cjs file.
- Enhanced the TypeScript configuration to include an additional source file for improved workspace management.
