+++ gnome-extension/.prettierrc
@@ -1,0 +1,10 @@
+{
+    "singleQuote": true,
+    "printWidth": 100,
+    "endOfLine": "lf",
+    "semi": true,
+    "tabWidth": 4,
+    "trailingComma": "es5",
+    "arrowParens": "avoid",
+    "quoteProps": "as-needed"
+}
