--- gnome-extension/.prettierrc.cjs
@@ -1,11 +1,0 @@
-/** @type {import('prettier').Config} */
-module.exports = {
-    singleQuote: true,
-    printWidth: 100,
-    endOfLine: 'lf',
-    semi: true,
-    tabWidth: 4,
-    trailingComma: 'es5',
-    arrowParens: 'avoid',
-    quoteProps: 'as-needed',
-};
