+++ gnome-extension/tsconfig.json
@@ -1,7 +1,9 @@
 {
     "compilerOptions": {
         "strict": true,
-        "lib": ["ES2022"],
+        "lib": [
+            "ES2022"
+        ],
         "target": "ES2022",
         "module": "NodeNext",
         "moduleResolution": "NodeNext",
@@ -14,6 +16,12 @@
         "forceConsistentCasingInFileNames": true,
         "resolveJsonModule": true
     },
-    "include": ["src/types/ambient.d.ts"],
-    "files": ["src/extension.ts", "src/prefs.ts"]
+    "include": [
+        "src/types/ambient.d.ts"
+    ],
+    "files": [
+        "src/extension.ts",
+        "src/prefs.ts",
+        "src/core.ts"
+    ]
 }
